fix: remove CollectiveCover
