+++ components/Hosts.js
@@ -1,11 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
+import { defineMessages, FormattedMessage, injectIntl } from 'react-intl';
+
 import Header from './Header';
 import Body from './Body';
 import Footer from './Footer';
 import HostsWithData from './HostsWithData';
-import HostsCover from './HostsCover';
-import { defineMessages, injectIntl } from 'react-intl';
+
+import Container from './Container';
+import { H1, P } from './Text';
+import Link from './Link';
+
+const CoverSmallCTA = styled.span`
+  a:hover {
+    text-decoration: underline !important;
+  }
+`;
 
 class Hosts extends React.Component {
   static propTypes = {
@@ -27,6 +38,10 @@ class Hosts extends React.Component {
         defaultMessage:
           "Hosts are legal entities that collect money on behalf of open collectives so that they don't have to worry about accounting, taxes, etc. Some also provide extra services. {findOutMoreLink}",
       },
+      'hosts.findOutMoreLink': {
+        id: 'hosts.description.findOutMoreLink',
+        defaultMessage: 'Find out more about becoming an Open Collective Host.',
+      },
     });
   }
 
@@ -34,9 +49,22 @@ class Hosts extends React.Component {
     const { LoggedInUser, intl } = this.props;
 
     const title = intl.formatMessage(this.messages['hosts.title']);
-    const description = intl.formatMessage(this.messages['hosts.description'], {
-      findOutMoreLink: '',
-    });
+
+    const findOutMoreMessage = intl.formatMessage(this.messages['hosts.findOutMoreLink']);
+
+    const findOutMoreLink = (
+      <CoverSmallCTA>
+        <Link route="https://docs.opencollective.com/help/hosts/become-host">{findOutMoreMessage}</Link>
+      </CoverSmallCTA>
+    );
+
+    const description = (
+      <FormattedMessage
+        id="hosts.description"
+        defaultMessage="Hosts are legal entities that collect money on behalf of open collectives so that they don't have to worry about accounting, taxes, etc. Some also provide extra services. {findOutMoreLink}"
+        values={{ findOutMoreLink }}
+      />
+    );
 
     return (
       <div className="Hosts">
@@ -67,7 +95,12 @@ class Hosts extends React.Component {
         />
 
         <Body>
-          <HostsCover title={title} description={description} href="/hosts" className="small" />
+          <Container mt={2} mb={2}>
+            <H1 fontSize={['H4', 'H2']} lineHeight={3} fontWeight="bold" textAlign="center" color="black.900">
+              {title}
+            </H1>
+            <P textAlign="center">{description}</P>
+          </Container>
 
           <div className="content">
             <HostsWithData LoggedInUser={LoggedInUser} />
